Parsefloat instead parseint

+++ cryptowallet/src/components/transactions/TransactionReceive.js
@@ -8,6 +8,6 @@ export default ({ exchangeInfo }) => {
         const { value, decimals, symbol } = exchangeInfo;
         return <span className="text-primary"><TransactionField valid={true}
                                  name="You receive"
-                                 value={`${Math.pow(10, -1 * parseInt(decimals)) * parseInt(value)} ${symbol}`}/></span>
+                                 value={`${Math.pow(10, -1 * parseInt(decimals)) * parseFloat(value)} ${symbol}`}/></span>
     }
 }
